Fix acceptance tests
In 2940279 management of the apache vhost was split into a separate
class and the vhost_name parameter was removed from the
bandersnatch::mirror class. This wasn't done in a backwards compatible
way and the acceptance tests were never updated to use the new API.
This patch updates the tests to use the new class and to stop using the
now invalid parameter.

Change-Id: If724c51b72f634a95bbffd080a9d33234a5d7645