Allow any user to run bandersnatch
Currently, bandersnatch defaults to the root user.
Configuring the crom job to a different user fails because the init class
sets the /var/log/bandersnatch and /var/run/bandersnatch
to the default user which causes permission denied issue.

Update the init.pp to allow setting the user and group owernship to the
same as the cron.pp and mirror.pp

Change-Id: I69ad9635f014ab522b27d996592ce56389061d9c