Pin bandersnatch to 1.11
Because the world regularly loses its collective mind when living under
the aegis of the god known as SemVer, clearly it's ok to break users when you
do it while bumping the major version.

Humble supplicants that we are, bowing before the mightiness and
righteous glory that places developer convenience and code cleanliness
over actual users actually using the actual software, we must prostrate
ourselves and beg for forgiveness while pinning the software to the
previous working version so that our users can have a mirror that works.

Woe betide us that we do not honor the obession with discarding
compatibility and rushing madly headlong into the abyss of the future.
Clearly we are not worthy to sit at the right hand of Tom the father and
his legion of cohorts, but instead must actually run a production
service.

Change-Id: If7bcf7cd93ac0e22d3294e401556f6642389f039